requirements.txt
streamlit
pandas
openai
qdrant-client
pypdf
python-dotenv
requests
